fix(usage): 面板/管理端用量接口的请求计数 cast 改 float
QuotaService 上一轮已改 float，但 panel/Usage.php 与 dashboard/Usage.php
控制器出口还有一批 (int) 把小数用量截成整数。token 相关保持 int，请求
用量（balance/used/chargedIn/delta/charge）改 (float)。

+++ app/controller/dashboard/Usage.php
@@ -78,7 +78,7 @@ class Usage extends BaseController
             $plan     = $p['billing_plan'];
             $byReq    = $plan === 'coding';
             $res      = $reservedMap[$plan] ?? ['tokens' => 0, 'requests' => 0];
-            $balance  = $byReq ? (int) $p['request_balance'] : (int) $p['token_balance'];
+            $balance  = $byReq ? (float) $p['request_balance'] : (int) $p['token_balance'];
             $reserve  = $byReq ? $res['requests'] : $res['tokens'];
             $plans[]  = [
                 'billingPlan' => $plan,
@@ -108,7 +108,7 @@ class Usage extends BaseController
                 'email'          => $r['email'],
                 'role'           => $r['role'],
                 'tokenBalance'   => (int) $r['token_balance'],
-                'requestBalance' => (int) $r['request_balance'],
+                'requestBalance' => (float) $r['request_balance'],
             ];
         }, $users);
